LF-5478 Rename utils file and remove unused export
